% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BackwardElim.R
\name{BackwardElim}
\alias{BackwardElim}
\title{BackwardElim.}
\usage{
BackwardElim(data, response, makeModels = TRUE, suppressOutput = FALSE)
}
\arguments{
\item{data}{a dataframe to be used for modelling.}

\item{response}{a string representing the name of the response
                variable (must exactly match that column name).}

\item{makeModels}{a boolean that determines whether or not this function
       leaves all the models it makes in your workspace or not.
       \code{TRUE} by default.}

\item{suppressOutput}{a boolean that you can use to stop the program
       from blowin' up yo console... \code{FALSE} by default.}
}
\value{
a list of linear models, all the best found while using
        backward elimination! Unless
}
\description{
A quick and dirty function for backward elimintation.
}
\details{
This function takes a data set and creates a series of linear models
using backward elimination with the intention of creating the best n
models where n is equal to the number of predictor variables.
}

