% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ForwardSelect.R
\name{ForwardSelect}
\alias{ForwardSelect}
\title{Forward Selection.}
\usage{
ForwardSelect(data, response, makeModels = TRUE, suppressOutput = FALSE)
}
\arguments{
\item{data}{a dataframe including the data you want to model.}

\item{response}{a string denoting which variable is the response.
                note: this must EXACTLY match the column's header.}

\item{makeModels}{a boolean that decides whether or to return all the
                  models in a list. \code{TRUE} by default.}

\item{suppressOutput}{another boolean that decides whether or not
                      this function will blow up your console with
                      knowledge... \code{FALSE} by default.}
}
\value{
a list of 'lm' objects, or nothing, depending on the value
        passed to the argument makeModels.
}
\description{
Just a quick and dirty forward selection algorithm.
}
\details{
This function just whips through your provided data and builds a series
of linear models selecting each new model from the previous using
forward selection. Nothing fancy.
}

