% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kFoldCV.R
\name{kFoldCV}
\alias{kFoldCV}
\title{K-Fold Cross Validation}
\usage{
kFoldCV(data, formula, k = 4)
}
\arguments{
\item{data}{a data frame containing the observations for modelling.}

\item{formula}{an R 'formula' object to determine the linear model
               to be constructed.}

\item{k}{the number of folds to use. Defaults to 4.}
}
\value{
a number, specifically the mean squared prediction error
        for the cross validated sub-samples made by the K-Fold
        Cross Validation algorithm.
}
\description{
K-Fold Cross Validation for Linear Models.
}
\details{
This is a really primitive funciton for calculating Mean Squared
Prediction Error (MSPE) for K-Fold Cross Validation of Linear
Models. Pretty rudimentary...
}

