% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeaveOneOutCV.R
\name{LeaveOneOutCV}
\alias{LeaveOneOutCV}
\title{Leave-One-Out Cross Validation.}
\usage{
LeaveOneOutCV(data, formula)
}
\arguments{
\item{data}{a dataframe containing the data you want to check your model.
            with.}

\item{formula}{a 'formula' object that can be used to build linear models.}
}
\value{
a number (the MSE of the L.O.O. C.V.)
}
\description{
Calculate Leave-One-Out Cross Validation Mean Squared Error.
}
\details{
This function just takes a data set and a linear model and calculates
the mean squared prediction error using the Leave-One-Out Cross Validation
algorithm.
}

